# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Assignment of projects API
  description: API
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api'
paths:
  /auth/login:
    get:
      summary: Авторизация
      security: []
      responses:
        '200':
          description: Страница авторизации

    post:
      summary: Авторизация
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInDto'
      responses:
        '200':
          description: Токен авторизации
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
        '400':
          description: Неверный формат email
        '401':
          description: Неверный логин или пароль

  /courses:
    get:
      summary: Список курсов
      security:
        - bearerAuth: []
      parameters:
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Список курсов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          description: Не авторизован

    post:
      summary: Создание курса
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseDto'
      responses:
        '201':
          description: Курс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на создание

  /courses/{id}:
    get:
      summary: Получение одного курса
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные курса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          description: Не авторизован
        '404':
          description: Курс не найден

    put:
      summary: Обновление курса
      security:
        - bearerAuth: []
      x-roles: [ 'admin' ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseDto'
      responses:
        '200':
          description: Курс обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на редактирование
        '404':
          description: Курс не найден

    delete:
      summary: Удаление курса
      security:
        - bearerAuth: []
      x-roles: [ 'admin' ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Курс удален
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на удаление
        '404':
          description: Курс не найден

  /courses/{id}/distribution:
    get:
      summary: Визуализация распределения по курсу
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Распределение по курсу
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionResult'
        '401':
          description: Не авторизован
        '404':
          description: Курс не найден

  /courses/{id}/projects:
    get:
      summary: Список проектов определенного курса
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: query
          schema: { type: integer }
        - name: title
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [title, createdAt]
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Не авторизован
        '404':
          description: Проекты курса не найдены

  /projects:
    get:
      summary: Список всех проектов
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: query
          schema: { type: integer }
        - name: teacherId
          in: query
          schema: { type: integer }
        - name: title
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [title, createdAt]
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Не авторизован

    post:
      summary: Создание проекта
      security:
        - bearerAuth: []
      x-roles: [ 'admin' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, teacherId, courseId]
              properties:
                title:
                  type: string
                teacherId:
                  type: integer
                courseId:
                  type: integer
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на создание
        '404':
          description: Курс или преподаватель не найдены

  /projects/{id}:
    get:
      summary: Получение проекта
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Не авторизован
        '404':
          description: Проект не найден

    put:
      summary: Обновление проекта
      security:
        - bearerAuth: []
      x-roles: [ 'admin' ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                teacherId: { type: integer }
                courseId: { type: integer }
      responses:
        '200':
          description: Проект обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Неверные данные
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на редактирование
        '404':
          description: Проект не найден

    delete:
      summary: Удаление проекта
      security:
        - bearerAuth: []
      x-roles: [ 'admin' ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Проект удален
        '401':
          description: Не авторизован
        '403':
          description: Нет прав на удаление
        '404':
          description: Проект не найден

  /assignments:
    get:
      summary: Общее распределение студентов по проектам
      parameters:
        - name: courseId
          in: query
          schema: { type: integer }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [assignedAt, createdAt]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Распределение
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assignment' }
        '401':
          description: Не авторизован

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignInDto:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    User:
      type: object
      required: [uid, firstName, lastName, email, role, ldapUid]
      properties:
        uid: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [student, admin]
        group: { type: integer }
        ldapUid: { type: string }
        createdAt: { type: string, format: date-time }
    Course:
      type: object
      required: [uid, name, maxTeamSize, minTeamSize, isActive, adminId]
      properties:
        uid: { type: integer }
        name: { type: string }
        semester: { type: integer }
        isActive: { type: boolean }
        maxTeamSize: { type: integer }
        minTeamSize: { type: integer }
        adminId: { type: integer }
        description: { type: string }
        registrationDeadline: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Project:
      type: object
      required: [uid, title, teacherId, teacherFirstName, teacherLastName, courseName, courseId]
      properties:
        uid: { type: integer }
        title: { type: string }
        teacherFirstName: { type: string }
        teacherLastName: { type: string }
        courseName: { type: string }
        courseId: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Assignment:
      type: object
      required: [studentId, studentFirstName, studentLastName, projectId, projectName, courseId, courseName]
      properties:
        studentId: { type: integer }
        studentFirstName: { type: string }
        studentLastName: { type: string }
        projectId: { type: integer }
        projectName: { type: string }
        courseId: { type: integer }
        courseName: { type: string }
        assignedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DistributionResult:
      type: object
      properties:
        course:
          $ref: '#/components/schemas/Course'
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/Assignment'
    CreateCourseDto:
      type: object
      required: [name, maxTeamSize, minTeamSize, isActive]
      properties:
        name: { type: string, minLength: 3 }
        maxTeamSize: { type: integer, minimum: 1 }
        minTeamSize: { type: integer, minimum: 1 }
        isActive: { type: boolean }
        registrationDeadline: { type: string, format: date-time }
    UpdateCourseDto:
      type: object
      properties:
        name: { type: string, minLength: 3 }
        maxTeamSize: { type: integer, minimum: 1 }
        minTeamSize: { type: integer, minimum: 1 }
        isActive: { type: boolean }
        registrationDeadline: { type: string, format: date-time }